Add support for hiding diff signs (leading + and -)
The `+` and `-` signs in diff views can be removed by setting the new
option `diff-hide-signs`. When set, only the color of the lines
distinguishes added, removed and context lines.

Empty added or removed lines are indicated by a space in column 0 with
the "diff highlight" colors which has the `reverse` property by default.

Closes #855